requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyMuPDF==1.22.5
python-dotenv==0.19.0
pydantic==2.5.1